requirements.txt
sanic==0.6.0
logzero==1.3.0
python-dotenv==0.7.1